Chap7) Securing

+++ planner/routes/events.py
@@ -1,6 +1,8 @@
 from beanie import PydanticObjectId
-from fastapi import APIRouter, Body, HTTPException, status
+from fastapi import APIRouter, Depends, HTTPException, status
 from database.connection import Database
+
+from auth.authenticate import authenticate
 
 from models.events import Event, EventUpdate
 from typing import List
@@ -30,19 +32,25 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
 
 
 @event_router.post("/new")
-async def create_event(body: Event) -> dict:
+async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {"message": "Event created successfully"}
 
 
 @event_router.delete("/{id}")
-async def delete_event(id: PydanticObjectId) -> dict:
-    event = await event_database.delete(id)
+async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
+    event = await event_database.get(id)
     if not event:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Event with supplied ID does not exist",
+            status_code=status.HTTP_404_NOT_FOUND, detail="Event not found"
         )
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST, detail="Operation not allowed"
+        )
+    event = await event_database.delete(id)
+
     return {"message": "Event deleted successfully."}
 
 
@@ -53,7 +61,14 @@ async def delete_all_events() -> dict:
 
 
 @event_router.put("/{id}", response_model=Event)
-async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
+async def update_event(
+    id: PydanticObjectId, body: EventUpdate, user: str = Depends(authenticate)
+) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST, detail="Operation not allowed"
+        )
     updated_event = await event_database.update(id, body)
     if not updated_event:
         raise HTTPException(
